backend/requirements.txt
fastapi
uvicorn
supabase
langchain-core>=1.2.5,<2.0.0
langchain-classic>=1.0.0,<2.0.0
langchain-openai>=0.1.0,<1.0.0
sse-starlette
pydantic>=2.7.0
pydantic-settings
python-dotenv
langchain-community==0.4.1
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0,<3.0.0
slowapi>=0.1.9
